requirements.txt
baidu-aip
wand
jieba
fire